requirements.txt
streamlit
sqlalchemy
requests
psycopg2-binary
python-dotenv
